requirements.txt
click==8.0.4
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scikit-surprise==1.1.1
scipy==1.9.0
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0